+++ gems/aws-sdk-s3/lib/aws-sdk-s3/plugins/iad_regional_endpoint.rb
@@ -17,7 +17,8 @@ region. Defaults to `legacy` mode using global endpoint.
         end
 
         def add_handlers(handlers, config)
-          if config.region == 'us-east-1'
+          # only modify non-custom endpoints
+          if config.regional_endpoint && config.region == 'us-east-1'
             handlers.add(Handler)
           end
         end
@@ -29,10 +30,8 @@ region. Defaults to `legacy` mode using global endpoint.
             # keep legacy global endpoint pattern by default
             if context.config.s3_us_east_1_regional_endpoint == 'legacy'
               host = context.http_request.endpoint.host
-              # if it's an ARN or private link endpoint, then
-              # don't touch the endpoint at all
-              unless context.metadata[:s3_arn] ||
-                     IADRegionalEndpoint.private_link_host?(host, context)
+              # if it's an ARN then don't touch the endpoint at all
+              unless context.metadata[:s3_arn]
                 legacy_host = IADRegionalEndpoint.legacy_host(host)
                 context.http_request.endpoint.host = legacy_host
               end
@@ -47,20 +46,6 @@ region. Defaults to `legacy` mode using global endpoint.
         end
 
         private
-
-        def self.private_link_host?(host, context)
-          force_path_style = context.config.force_path_style
-          private_link_endpoint_prefix = 'bucket'.freeze
-
-          if force_path_style
-            # bucket is in the path, so use first member
-            actual_endpoint_prefix = host.split(".")[0]
-          else
-            # bucket is a host label, so use second member
-            actual_endpoint_prefix = host.split(".")[1]
-          end
-          private_link_endpoint_prefix == actual_endpoint_prefix
-        end
 
         def self.resolve_iad_regional_endpoint(cfg)
           mode = ENV['AWS_S3_US_EAST_1_REGIONAL_ENDPOINT'] ||