+++ gems/aws-sdk-s3/CHANGELOG.md
@@ -1,7 +1,7 @@
 Unreleased Changes
 ------------------
 
-* Issue - Fix an issue with the IAD regional endpoint plugin removing `us-east-1` from a PrivateLink endpoint.
+* Issue - Fix an issue with the IAD regional endpoint plugin removing `us-east-1` from custom endpoints.
 
 1.88.1 (2021-02-12)
 ------------------